requirements.txt
Flask==2.3.3
requests==2.31.0
schedule==1.1.0
google-api-python-client==2.94.0
python-telegram-bot==20.3
